% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_meta.R
\name{attach_meta_coord}
\alias{attach_meta_coord}
\title{Anexar: Coordenadas geográficas}
\usage{
attach_meta_coord(df, sheetname = "map_us_coord")
}
\arguments{
\item{df}{Dataframe já transformado através de \code{process_sisma_csv} ou \code{process_sisma_export}}

\item{sheetname}{Nome da folha do MISAU_Site_Map}
}
\value{
Dataframe contendo coordenadas geográficas das unidades sanitárias
}
\description{
\code{attach_meta_coord} Une as coordenadas geográficas da unidade sanitária gravadas no pacote \code{sismar} ao objeto de dados introduzido pelo utilizador
}
\examples{
 \dontrun{

 df <- attach_meta_uid(df)}
}
