% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_meta.R
\name{attach_meta_mq}
\alias{attach_meta_mq}
\title{Retreive and attach MQ metadata}
\usage{
attach_meta_mq(df)
}
\arguments{
\item{df}{Dataframe cleaned via reshape_sisma}
}
\value{
A dataframe containing mq metadata
}
\description{
\code{attach_meta_mq} utilizes stored user credentials to access
a GoogleSheet where province and district uid metadata is
stored.
}
\examples{
 \dontrun{

 df <- attach_meta_mq(df)}
}
