% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_meta.R
\name{attach_meta_uid}
\alias{attach_meta_uid}
\title{Anexar: Metadados uid das provincias e distritos}
\usage{
attach_meta_uid(df, sheetname = "map_distrito_psnuuid")
}
\arguments{
\item{df}{Quadro de dados limpo através de \code{process_sisma_export}}

\item{sheetname}{Nome da folha do MISAU_Site_Map}
}
\value{
Quadro de dados contendo colunas para metadados das provincias e distritos
}
\description{
\code{attach_meta_uid} Une informações dos UIDs das provincias e distritos
}
\examples{
 \dontrun{

 df <- attach_meta_uid(df)}
}
