% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_exact_match.R
\name{check_exact_match}
\alias{check_exact_match}
\title{Helper function to check for input matches against template}
\usage{
check_exact_match(df_vars, template_vars)
}
\arguments{
\item{df_vars}{dataframe variables}

\item{template_vars}{reference input variables}
}
\value{
matrix of match vs. no match
}
\description{
Helper function to check for input matches against template
}
\examples{
 \dontrun{

 df <- check_exact_match()}
}
