% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_disa_dpi.R
\name{clean_disa_dpi}
\alias{clean_disa_dpi}
\title{Helper function for recoding DISA dataframe}
\usage{
clean_disa_dpi(df)
}
\arguments{
\item{df}{Semi-processed dataframe passed in via process_disa_dpi}
}
\value{
Recoded DISA dataframe
}
\description{
Helper function for recoding DISA dataframe
}
\examples{
 \dontrun{

 df <- clean_disa_dpi()}
}
