% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_sisma_csv.R
\name{clean_sisma_csv}
\alias{clean_sisma_csv}
\title{Helper function for cleaning a .csv pivot table export from SISMA}
\usage{
clean_sisma_csv(file)
}
\arguments{
\item{file}{A .csv file downloaded from SISMA with data element or indicator results running wide}
}
\value{
A data frame object in long format with 7 variables (period, snu, psnu, sitename, sisma_uid, indicator, value)
}
\description{
Helper function for cleaning a .csv pivot table export from SISMA
}
\examples{
 \dontrun{

 df <- clean_sisma_csv()}
}
