% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_sisma.R
\name{clean_sisma_df}
\alias{clean_sisma_df}
\title{Limpar ficheiro bruto SISMA}
\usage{
clean_sisma_df(df)
}
\arguments{
\item{df}{um objecto de dados importado através de \code{process_sisma_export}}
}
\value{
\code{clean_sisma_df()} devolve um objeto de dados limpo com 7 colunas (período, província, distrito, us, sisma_uid, indicador e valor)
}
\description{
\code{clean_sisma_df()} pega num objecto de dados importado através de \code{process_sisma_export} e devolve-o transformado numa estrutura longa, cria uma coluna de período formatada e devolve um objeto de dados limpo com 7 colunas
}
\examples{
 \dontrun{

 df <- clean_sisma_df()}
}
