% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_report.R
\name{create_report}
\alias{create_report}
\title{Create a dataset for completeness check}
\usage{
create_report(data_delivered, sites)
}
\arguments{
\item{data_delivered}{Data received from all sites}

\item{sites}{sites that should deliver data}
}
\value{
a tibble with expected vs actual data submitted at a site level
}
\description{
Create a dataset for completeness check
}
\examples{
 \dontrun{
   create_report(data_delivered, sites)
 }

}
