% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_disa_uid_map}
\alias{data_disa_uid_map}
\title{UID map for linking DISA, SISMA and DATIM}
\format{
A list of the variables included in the site list
\describe{
\item{data_disa_uid_map}{disa_uid}
\item{data_disa_uid_map}{sisma_uid}
\item{data_disa_uid_map}{datim_uid}
}
}
\usage{
data(data_disa_uid_map)
}
\description{
List of indicators
}
\keyword{datasets}
