% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_sisma_ats_ci_map}
\alias{data_sisma_ats_ci_map}
\title{Variable map for engineering features specific to SISMA ATS Case Index/Linkage dataframe}
\format{
A list of the variables included to engineer dataset features
\describe{
\item{data_sisma_ats_ci_map}{indicator}
\item{data_sisma_ats_ci_map}{indicator_new}
\item{data_sisma_ats_ci_map}{disaggregate}
\item{data_sisma_ats_ci_map}{sub_group}
\item{data_sisma_ats_ci_map}{result_status}
\item{data_sisma_ats_ci_map}{age_coarse}
}
}
\usage{
data(data_sisma_ats_ci_map)
}
\description{
List of indicators
}
\keyword{datasets}
