% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_sisma_ats_smi_map}
\alias{data_sisma_ats_smi_map}
\title{Variable map for engineering features specific to SISMA ATS SMI dataframe}
\format{
A list of the variables included to engineer dataset features
\describe{
\item{data_sisma_ats_smi_map}{indicator}
\item{data_sisma_ats_smi_map}{indicator_new}
\item{data_sisma_ats_smi_map}{modality}
\item{data_sisma_ats_smi_map}{sub_group}
\item{data_sisma_ats_smi_map}{result_status}
\item{data_sisma_ats_smi_map}{age_coarse}
\item{data_sisma_ats_smi_map}{sex}
}
}
\usage{
data(data_sisma_ats_smi_map)
}
\description{
List of indicators
}
\keyword{datasets}
