% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_sisma_geo_above_site}
\alias{data_sisma_geo_above_site}
\title{Map for linking province and district shape codes to SISMA data}
\format{
A list of the variables included in the site list
\describe{
\item{data_sisma_geo_above_site}{provincia}
\item{data_sisma_geo_above_site}{snuuid}
\item{data_sisma_geo_above_site}{distrito}
\item{data_sisma_geo_above_site}{psnuuid}
}
}
\usage{
data(data_sisma_geo_above_site)
}
\description{
List of indicators
}
\keyword{datasets}
