% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_sisma_geo_sites}
\alias{data_sisma_geo_sites}
\title{Mapa para ligar as coordenadas geográficas aos dados das unidades de saúde SISMA}
\format{
Uma lista das variáveis incluídas no objecto de dados
\describe{
\item{sisma_uid}{UID SISMA (e.g. x9jhds3ie6)}
\item{latitude}{Latitude geográfica}
\item{longitude}{Longitude geográfica}
}
}
\usage{
data(data_sisma_geo_sites)
}
\description{
Mapa para ligar as coordenadas geográficas aos dados das unidades de saúde SISMA
}
\keyword{datasets}
