% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_sisma_hiv_ajmhcmm_map}
\alias{data_sisma_hiv_ajmhcmm_map}
\title{Variable map for engineering features specific to SISMA AJM/HC/MM dataframe}
\format{
A list of the variables included to engineer dataset features
\describe{
\item{data_sisma_hiv_ajmhcmm_map}{indicator}
\item{data_sisma_hiv_ajmhcmm_map}{indicator_new}
\item{data_sisma_hiv_ajmhcmm_map}{age}
\item{data_sisma_hiv_ajmhcmm_map}{sex}
\item{data_sisma_hiv_ajmhcmm_map}{result_status}
\item{data_sisma_hiv_ajmhcmm_map}{disaggregate}
\item{data_sisma_hiv_ajmhcmm_map}{source}
}
}
\usage{
data(data_sisma_hiv_ajmhcmm_map)
}
\description{
List of indicators
}
\keyword{datasets}
