% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_sisma_shape_map}
\alias{data_sisma_shape_map}
\title{Map for linking province and district shape codes to SISMA data}
\format{
A list of the variables included in the site list
\describe{
\item{data_sisma_shape_map}{provincia}
\item{data_sisma_shape_map}{snuuid}
\item{data_sisma_shape_map}{distrito}
\item{data_sisma_shape_map}{psnuuid}
}
}
\usage{
data(data_sisma_shape_map)
}
\description{
List of indicators
}
\keyword{datasets}
