% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_sisma_sitelist}
\alias{data_sisma_sitelist}
\title{Lista de unidades sanitárias SISMA}
\format{
Uma lista das variáveis incluídas no objecto de dados
\describe{
\item{sisma_uid}{UID SISMA (e.g. x9jhds3ie6)}
\item{provincia}{Provincia}
\item{distrito}{Distrito}
\item{us}{Unidade Sanitária}
}
}
\usage{
data(data_sisma_sitelist)
}
\description{
Lista de unidades sanitárias SISMA
}
\keyword{datasets}
