% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_sisma_smi_ccr_map}
\alias{data_sisma_smi_ccr_map}
\title{SISMA HIV SISMA CCR Map for Feature Engineering}
\format{
A list of the variables included to engineer dataset features
\describe{
\item{data_sisma_smi_ccr_map}{indicator}
\item{data_sisma_smi_ccr_map}{indicator_new}
\item{data_sisma_smi_ccr_map}{disaggregate}
\item{data_sisma_smi_ccr_map}{source}
}
}
\usage{
data(data_sisma_smi_ccr_map)
}
\description{
List of indicators
}
\keyword{datasets}
