% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_sisma_smi_cpp_map}
\alias{data_sisma_smi_cpp_map}
\title{SISMA Consulta Post-Parto Map for Feature Engineering}
\format{
A list of the variables included to engineer dataset features
\describe{
\item{data_sisma_smi_cpp_map}{indicator}
\item{data_sisma_smi_cpp_map}{indicator_new}
\item{data_sisma_smi_cpp_map}{sex}
\item{data_sisma_smi_cpp_map}{age}
\item{data_sisma_smi_cpp_map}{sub_group}
\item{data_sisma_smi_cpp_map}{age_coarse}
\item{data_sisma_smi_cpp_map}{result_status}
\item{data_sisma_smi_cpp_map}{disaggregate}
\item{data_sisma_smi_cpp_map}{disaggregate_sub}
\item{data_sisma_smi_cpp_map}{source}
}
}
\usage{
data(data_sisma_smi_cpp_map)
}
\description{
List of indicators
}
\keyword{datasets}
