% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_sisma_smi_pf_map}
\alias{data_sisma_smi_pf_map}
\title{Variable map for engineering features specific to SISMA SMI PF dataframe}
\format{
A list of the variables included to engineer dataset features
\describe{
\item{data_sisma_smi_pf_map}{indicator}
\item{data_sisma_smi_pf_map}{indicator_new}
\item{data_sisma_smi_pf_map}{sex}
\item{data_sisma_smi_pf_map}{age}
\item{data_sisma_smi_pf_map}{disaggregate}
\item{data_sisma_smi_pf_map}{disaggregate_sub}
\item{data_sisma_smi_pf_map}{source}
}
}
\usage{
data(data_sisma_smi_pf_map)
}
\description{
List of indicators
}
\keyword{datasets}
