% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedup_naomi.R
\name{dedup_naomi}
\alias{dedup_naomi}
\title{Deduplicate Naomi CSV file}
\usage{
dedup_naomi(
  filename,
  quarter_lab = c("September 2023"),
  area_level_lab = c("Distrito")
)
}
\arguments{
\item{filename}{Path to Naomi CSV output}

\item{quarter_lab}{Define quarter for Naomi output}

\item{area_level_lab}{Define geography level for Naomi output}
}
\value{
A dataframe containing deduplicated Naomi indicator results
}
\description{
\code{dedup_naomi} Deduplicates Naomi CSV file and maps in snu1uid
and psnuuid so that modeled indicator results can be plotted
geographically more easily
}
\examples{
 \dontrun{

 df <- dedup_naomi(df)}
}
