% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_site_table.R
\name{generate_site_table}
\alias{generate_site_table}
\title{Generate cleaned health facility table}
\usage{
generate_site_table(file)
}
\arguments{
\item{file}{Path to a .csv file exported from SISMA that contains
a full list of health facilities}
}
\value{
\code{generate_site_table} returns a object with 7 deduplicated columns (sisma_uid,
snu, psnu, sitename)
}
\description{
Generate cleaned health facility table
}
\examples{
 \dontrun{
 df <- generate_site_table()}
}
