% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_check_disa.R
\name{id_check_disa}
\alias{id_check_disa}
\title{Check for DISA UID's without corresponding SISMA UID's and for SISMA UID's with multiple DISA UID's}
\usage{
id_check_disa(df, period_window = 6)
}
\arguments{
\item{df}{Dataframe already processed by either \code{process_disa_cv} or \code{process_disa_dpi}}

\item{period_window}{Number of months returned in the output (counted from the most current month in the input dataframe)}
}
\value{
Dataframe with multiple DISA UID's mapping to a SISMA UID or DISA UID's without a corresponding SISMA UID
}
\description{
Check for DISA UID's without corresponding SISMA UID's and for SISMA UID's with multiple DISA UID's
}
\examples{
 \dontrun{

 df <- id_check_disa()}
}
