% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pop.R
\name{load_pop_ine}
\alias{load_pop_ine}
\title{Importar quadro de dados populacionais INE}
\usage{
load_pop_ine(filename, input_tabs)
}
\arguments{
\item{filename}{utiliza um caminho fornecido pelo utilizador}

\item{input_tabs}{anos no ficheiro a ser processado}
}
\value{
um quadro contendo dados para os anos específicos
}
\description{
Importar quadro de dados populacionais INE
}
\examples{
 \dontrun{

 df <- load_pop_ine()}
}
