% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sisma_ats_ccsd.R
\name{parse_sisma_ats_ccsd}
\alias{parse_sisma_ats_ccsd}
\title{Create tidy dataframes for well and sick-child HIV testing}
\usage{
parse_sisma_ats_ccsd(df)
}
\arguments{
\item{df}{A dataframe object output from \code{clean_sisma_csv}}
}
\value{
\code{parse_sisma_ats_ccsd} returns a tidy object with 16 columns of
site metadata, indicator features and results
}
\description{
\code{parse_sisma_ats_ccsd} produces a tidy dataframe from an object passed in by
\code{sisma_clean_csv}. It engineers useful data features such as sex, age,
indicator disaggregation, sub-group type, etc.
}
\examples{
 \dontrun{

 df <- parse_sisma_ats_ccsd()}
}
