% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sisma_ats_index.R
\name{parse_sisma_ats_index}
\alias{parse_sisma_ats_index}
\title{A specific helper function for parsing a cleaned CSV export from SISMA}
\usage{
parse_sisma_ats_index(file)
}
\arguments{
\item{file}{Dataframe cleaned via reshape_sisma}
}
\value{
A tidy format of SISMA dataframe
}
\description{
A specific helper function for parsing a cleaned CSV export from SISMA
}
\examples{
 \dontrun{

 df <- parse_sisma_ats_index()}
}
