% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sisma_ats_results.R
\name{parse_sisma_ats_results}
\alias{parse_sisma_ats_results}
\title{A helper function for parsing a cleaned CSV export of ATS Results from SISMA}
\usage{
parse_sisma_ats_results(file)
}
\arguments{
\item{file}{Dataframe cleaned via reshape_sisma}
}
\value{
A tidy dataframe of SISMA ATS Results
}
\description{
A helper function for parsing a cleaned CSV export of ATS Results from SISMA
}
\examples{
 \dontrun{

 df <- parse_sisma_ats_results()}
}
