% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sisma_ats_saaj_cm.R
\name{parse_sisma_ats_saaj_cm}
\alias{parse_sisma_ats_saaj_cm}
\title{Create tidy dataframes for Youth-Friendly HIV testing}
\usage{
parse_sisma_ats_saaj_cm(df)
}
\arguments{
\item{df}{A dataframe object output from \code{clean_sisma_csv}}
}
\value{
\code{parse_sisma_ats_saaj_cm} returns a tidy object with 16 columns of
site metadata, indicator features and results
}
\description{
\code{parse_sisma_ats_saaj_cm} produces a tidy dataframe from an object passed in by
\code{sisma_clean_csv}. It engineers useful data features such as sex, age,
indicator disaggregation, sub-group type, etc.
}
\examples{
 \dontrun{

 df <- parse_sisma_ats_saaj_cm()}
}
