% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sisma_hiv_apss.R
\name{parse_sisma_hiv_apss}
\alias{parse_sisma_hiv_apss}
\title{Create tidy dataframes for HIV Psycho-social services}
\usage{
parse_sisma_hiv_apss(df)
}
\arguments{
\item{df}{A dataframe object output from \code{clean_sisma_csv}}
}
\value{
\code{parse_sisma_hiv_apss} returns a tidy object with 16 columns of
site metadata, indicator features and results
}
\description{
\code{parse_sisma_hiv_apss} produces a tidy dataframe from an object passed in by
\code{sisma_clean_csv}. It engineers useful data features such as sex, age,
indicator disaggregation, sub-group type, etc.
}
\examples{
 \dontrun{

 df <- parse_sisma_hiv_apss()}
}
