% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sisma_hiv_prep.R
\name{parse_sisma_hiv_prep}
\alias{parse_sisma_hiv_prep}
\title{A specific helper function for parsing a cleaned CSV export from SISMA}
\usage{
parse_sisma_hiv_prep(file)
}
\arguments{
\item{file}{Dataframe cleaned via reshape_sisma}
}
\value{
A tidy format of SISMA prep dataframe
}
\description{
A specific helper function for parsing a cleaned CSV export from SISMA
}
\examples{
 \dontrun{

 df <- parse_sisma_hiv_prep()}
}
