% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sisma_smi_ccr.R
\name{parse_sisma_smi_ccr}
\alias{parse_sisma_smi_ccr}
\title{A specific helper function for parsing a cleaned CSV export from SISMA}
\usage{
parse_sisma_smi_ccr(file)
}
\arguments{
\item{file}{Dataframe cleaned via reshape_sisma}
}
\value{
A tidy format of SISMA CCR dataframe
}
\description{
A specific helper function for parsing a cleaned CSV export from SISMA
}
\examples{
 \dontrun{

 df <- parse_sisma_smi_ccr()}
}
