% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sisma_smi_pf.R
\name{parse_sisma_smi_pf}
\alias{parse_sisma_smi_pf}
\title{Arrumar quadros de dados para os serviços do Planeamento Familiar (PF)}
\usage{
parse_sisma_smi_pf(df)
}
\arguments{
\item{df}{Objecto relatório padrão introduzido por \code{process_sisma_export}
ou \code{process_sisma_csv}}
}
\value{
\code{parse_sisma_smi_pf} devolve um quadro de dados arrumado com 16
colunas de metadados da us, desagregações do indicador, e valores
}
\description{
\code{parse_sisma_smi_pf} Uma função auxiliar que  gera um quadro de dados arrumado
a partir  de um objecto relatório padrão introduzido por \code{process_sisma_export}
ou \code{process_sisma_csv}. Variáveis úteis que são criados por esta função incluem
sexo, idade, tipo de subgrupo, e outras desagregações presentes na fonte.
}
\examples{
 \dontrun{

 df <- parse_sisma_smi_pf()}
}
