% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_disa_dpi.R
\name{process_disa_dpi}
\alias{process_disa_dpi}
\title{Processar: Diagnóstico precoce infantil (DPI) da DISA OpenLDR}
\usage{
process_disa_dpi(file, period, type = "new")
}
\arguments{
\item{file}{Caminho do ficheiro DPI .xlsx}

\item{period}{Mês dos dados DPI (formatado “AAAA-MM-DD”)}

\item{type}{Tipo de ficheiro de dados DPI a arrumar (“new” ou “old”)}
}
\value{
Um quadro de dados arrumado dos resultados do DPI
}
\description{
Processar: Diagnóstico precoce infantil (DPI) da DISA OpenLDR
}
\examples{
 \dontrun{

 df <- process_disa_dpi()}
}
