% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_naomi.R
\name{process_naomi}
\alias{process_naomi}
\title{Processar: Naomi 'indicators' .csv}
\usage{
process_naomi(
  filename,
  quarter_lab = c("September 2023"),
  area_level_lab = c("Distrito")
)
}
\arguments{
\item{filename}{Caminho para Naomi .csv}

\item{quarter_lab}{Define o trimestre para o qual as estimativas dos indicadores serão extraídas}

\item{area_level_lab}{Define o nível geográfico das estimativas geradas}
}
\value{
\code{process_naomi} devolve um objeto de dados que contém resultados do indicador Naomi deduplicados
}
\description{
\code{process_naomi} processa ficheiros .csv da Naomi e junta metadados para facilitar mapeamento geoespacial de indicadores
}
\examples{
 \dontrun{

 df <- process_naomi(df)}
}
