% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pop.R
\name{process_pop_ine}
\alias{process_pop_ine}
\title{Processar: Projecções demográficas distritais do INE}
\usage{
process_pop_ine(
  file_inventory,
  input_sheets,
  age_level = "Exact",
  output_type = "MISAU"
)
}
\arguments{
\item{file_inventory}{vector de ficheiros populacionais do INE em formato .xlsx}

\item{input_sheets}{vector (em caracteres) dos anos do ficheiro para inclusão no processamento}

\item{age_level}{Nível de pormenor da idade no quadro de dados arrumado (“Exato” por defeito com opção para “Agrupado” para faixas etárias de 5 anos)}

\item{output_type}{formato do provincia (por defeito "MISAU")}
}
\value{
\code{process_pop_ine} devolve um quadro de dados arrumado com 9 colunas
}
\description{
\code{process_pop_ine} usa um vector de ficheiros contendo projecções demográficas distritais do INE processando eles num único ficheiro arrumado
}
\examples{
 \dontrun{

 df <- process_pop_ine()}
}
