% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_sisma_csv.R
\name{process_sisma_csv}
\alias{process_sisma_csv}
\title{Function that cleans sisma csv exports and parses them into tidy dataframes}
\usage{
process_sisma_csv(file, type, language = "portuguese")
}
\arguments{
\item{file}{Path of sisma csv input}

\item{type}{Type of sisma csv export (CPN, ATS-R, etc.)}

\item{language}{Language of variable names (Portuguese or English)}
}
\value{
A tidy data frame of sisma program results
}
\description{
process_sisma_csv cleans and transforms programmatic
data sets exported in CSV format from SISMA into tidy data frames.
Such structured data sets are key to facilitating subsequent analysis
in R/PowerBI/Tableau.
\subsection{Overview of programmatic data type options}{
\itemize{
\item SMI-CPN: CPN, Saude Materna Infantil
\item SMI-MAT: Maternidade, Saude Materna Infantil
\item SMI-CCR: CCR, Saude Materna Infantil
\item ATS Result: ATS Resultados, Programa de HIV
\item ATS History: ATS Historial, Programa de HIV
\item ATS CI: ATS Caso Indice, Programa de HIV
}
}
}
\examples{
 \dontrun{

 df <- process_sisma_csv()}
}
