% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_sisma_export.R
\name{process_sisma_export}
\alias{process_sisma_export}
\title{Create tidy dataframes for standard SISMA .csv exports}
\usage{
process_sisma_export(filename, language = "portuguese")
}
\arguments{
\item{filename}{Path of sisma csv input}

\item{language}{Language of output variable names (portuguese or english)}
}
\value{
A tidy dataframe of SISMA program results
}
\description{
Create tidy dataframes for standard SISMA .csv exports
}
\examples{
 \dontrun{

 df <- process_sisma_export()}
}
