% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_example_csv.R
\name{read_example_csv}
\alias{read_example_csv}
\title{Lê dados de exemplo no ambiente R}
\usage{
read_example_csv(example_data)
}
\arguments{
\item{example_data}{Nome dos dados de exemplo gravado no \code{sismar}}
}
\value{
\code{read_example_csv} Quadro de dados de exemplo
}
\description{
Lê dados de exemplo no ambiente R
}
\examples{
 \dontrun{

 df <- example_data()}
}
