% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pop.R
\name{recode_ine_age}
\alias{recode_ine_age}
\title{Recodificar a idade em faixas etárias de cinco anos}
\usage{
recode_ine_age(df)
}
\arguments{
\item{df}{}
}
\value{
df
}
\description{
Recodificar a idade em faixas etárias de cinco anos
}
\examples{
 \dontrun{

 df <- recode_ine_age()}
}
