% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_create.R
\name{report_create}
\alias{report_create}
\title{Create a dataset for completeness check}
\usage{
report_create(data, report_type)
}
\arguments{
\item{data}{Data received from all sites}

\item{report_type}{sites that should deliver data}
}
\value{
a tibble with expected vs actual data submitted at a site level
}
\description{
Create a dataset for completeness check
}
\examples{
 \dontrun{
   report_create(data_delivered, sites)
 }

}
