% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_create_all.R
\name{report_create_all}
\alias{report_create_all}
\title{Create a report for all datasets in a specifical local folder.}
\usage{
report_create_all(file, period_selected)
}
\arguments{
\item{file}{dataset as a tibble}

\item{period_selected}{relevant month - "2023-01"}
}
\value{
a tibble with expected and actual results
}
\description{
Create a report for all datasets in a specifical local folder.
}
\examples{
 \dontrun{
   report_create_all(file, period_selected)
 }

}
