% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_write.R
\name{report_write}
\alias{report_write}
\title{Print out SNU, PSNU and Site detailes to different tabs in an excel sheet}
\usage{
report_write(data, file)
}
\arguments{
\item{data}{completeness check dataset}

\item{file}{path and filename to save to disk}
}
\value{
excel workbook with 3 tabs - SNU, PSNU and Site
}
\description{
Print out SNU, PSNU and Site detailes to different tabs in an excel sheet
}
\examples{
 \dontrun{
   report_write(data, file)
 }

}
