% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sismar_app.R
\name{run_sismar_app}
\alias{run_sismar_app}
\title{Launhc ShinyApp}
\usage{
run_sismar_app()
}
\value{
running shinyapp
}
\description{
Launhc ShinyApp
}
\examples{
 \dontrun{

 df <- run_sismar_app()}
}
