% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_language.R
\name{set_language}
\alias{set_language}
\title{Set language of output variables}
\usage{
set_language(df, language = "portuguese")
}
\arguments{
\item{df}{input dataframe}

\item{language}{variable output language}
}
\value{
dataframe with desired variable language
}
\description{
Set language of output variables
}
\examples{
 \dontrun{

 df <- set_language()}
}
