% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_sismar_folder.R
\name{setup_sismar_folder}
\alias{setup_sismar_folder}
\title{Configure project folder hierarchy}
\usage{
setup_sismar_folder(
  folder_list = list("Data", "Data/sismar", "Data/disa", "Data/pepfar",
    "Data/sismar/processed", "Data/disa/processed", "Data/pepfar/processed", "Images",
    "Scripts", "AI", "Dataout", "Data_public", "GIS", "Documents", "Graphics",
    "markdown")
)
}
\arguments{
\item{folder_list}{A list of desired folders}
}
\value{
A standard set of in-project folders to use with \code{sismar} functions
}
\description{
Configure project folder hierarchy
}
\examples{
 \dontrun{

 df <- setup_sismar_folder()}
}
