% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sismar-package.R
\docType{package}
\name{sismar-package}
\alias{sismar}
\alias{sismar-package}
\title{sismar: Arrumar dados SISMA}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Fornece um conjunto de funções para a criação de conjuntos de dados analíticos a partir de downloads do SISMA e DISA. Inclui funções que arrumam os ficheiros para um formato longo, removem variáveis desnecessárias, e criam colunas úteis para a análise.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://usaid-mozambique.github.io/sismar/}
  \item Report bugs at \url{https://github.com/usaid-mozambique/sismar/issues}
}

}
\author{
\strong{Maintainer}: Joe Lara \email{jlara@usaid.gov}

Authors:
\itemize{
  \item Emilton Cumbana \email{e.cumbana@outlook.com}
  \item Karishma Srikanth \email{ksrikanth@usaid.gov}
  \item Kirstin Lyon \email{kirstin.c.lyon@gmail.com}
}

}
\keyword{internal}
