% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_report.R
\name{write_report}
\alias{write_report}
\title{Print out SNU, PSNU and Site detailes to different tabs in an excel sheet}
\usage{
write_report(data, file)
}
\arguments{
\item{data}{completeness check dataset}

\item{file}{path and filename to save to disk}
}
\value{
excel workbook with 3 tabs - SNU, PSNU and Site
}
\description{
Print out SNU, PSNU and Site detailes to different tabs in an excel sheet
}
\examples{
 \dontrun{
   write_report(data, file)
 }

}
