% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_missing_vars.R
\name{add_missing_vars}
\alias{add_missing_vars}
\title{Função auxiliar que adiciona variáveis em falta nos quadros arrumados pelas funções \code{parse}}
\usage{
add_missing_vars(df)
}
\arguments{
\item{df}{Objecto quadro de dados processado pela família de funções \code{parse}}
}
\value{
Objecto quadro de dados contendo todos variáveis padrão
}
\description{
Função auxiliar que adiciona variáveis em falta nos quadros arrumados pelas funções \code{parse}
}
\examples{
 \dontrun{

 df <- add_missing_vars()}
}
