% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_meta_coord.R
\name{attach_meta_coord}
\alias{attach_meta_coord}
\title{Retreive and attach health facility geocoordinates metadata}
\usage{
attach_meta_coord(df, sheetname = "map_us_coord")
}
\arguments{
\item{df}{Dataframe cleaned via reshape_sisma}

\item{sheetname}{Sheet name from the MISAU_Site_Map}
}
\value{
A dataframe containing site metadata
}
\description{
\code{attach_meta_coord} utilizes stored user credentials to access
a GoogleSheet where health facility geocoordinates are
stored.
}
\examples{
 \dontrun{

 df <- attach_meta_uid(df)}
}
