% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_meta.R
\name{attach_meta_coord}
\alias{attach_meta_coord}
\title{Anexar coordenadas geográficas das unidades sanitárias}
\usage{
attach_meta_coord(df, sheetname = "map_us_coord")
}
\arguments{
\item{df}{Dataframe já transformado através de \code{process_sisma_csv} ou \code{process_sisma_export}}

\item{sheetname}{Nome da folha do MISAU_Site_Map}
}
\value{
Dataframe contendo coordenadas geográficas das unidades sanitárias
}
\description{
\code{attach_meta_coord} utiliza as credenciais de utilizador armazenadas para aceder
a uma folha de cálculo do Google onde as coordenadas geográficas das unidades de
saúde estão armazenadas.
}
\examples{
 \dontrun{

 df <- attach_meta_uid(df)}
}
