% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_meta.R
\name{attach_meta_mq}
\alias{attach_meta_mq}
\title{Anexar metadados melhoria da qualidade}
\usage{
attach_meta_mq(df)
}
\arguments{
\item{df}{Dataframe limpo através de \code{process_sisma_export}}
}
\value{
Um objeto de dados que contém metadados MQ
}
\description{
\code{attach_meta_mq} Une iniciativas de melhoria da qualidade gravadas no pacote \code{sismar} ao objeto de dados introduzido pelo utilizador
}
\examples{
 \dontrun{

 df <- attach_meta_mq(df)}
}
