% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_meta.R
\name{attach_meta_mq}
\alias{attach_meta_mq}
\title{Anexar metadados MQ}
\usage{
attach_meta_mq(df)
}
\arguments{
\item{df}{Dataframe limpo através de \code{process_sisma_export}}
}
\value{
Um objeto de dados que contém metadados MQ
}
\description{
\code{attach_meta_mq} utiliza credenciais de utilizador armazenadas para aceder a uma folha de cálculo do Google onde estão armazenados metadados de uid de província e distrito
}
\examples{
 \dontrun{

 df <- attach_meta_mq(df)}
}
