% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_meta.R
\name{attach_meta_partner}
\alias{attach_meta_partner}
\title{Retreive and attach partner metadata}
\usage{
attach_meta_partner(df, sheetname = "Alcancar")
}
\arguments{
\item{df}{Dataframe cleaned via reshape_sisma}

\item{sheetname}{Sheet name from the PCMD_Site_Map}
}
\value{
A dataframe containing partner metadata
}
\description{
\code{attach_meta_partner} utilizes stored user credentials to access
a GoogleSheet where province and district uid metadata is
stored.
}
\examples{
 \dontrun{

 df <- attach_meta_partner(df)}
}
