% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_meta.R
\name{attach_meta_partner}
\alias{attach_meta_partner}
\title{Anexar metadados do apoio dos parceiros}
\usage{
attach_meta_partner(df, sheetname = "Alcancar")
}
\arguments{
\item{df}{Dataframe limpo através de \code{process_sisma_export}}

\item{sheetname}{Nome da folha do PCMD_Site_Map}
}
\value{
A dataframe containing partner metadata
}
\description{
\code{attach_meta_partner} utiliza credenciais de utilizador armazenadas para aceder a uma folha de cálculo do Google onde estão armazenados metadados de uid de província e distrito
}
\examples{
 \dontrun{

 df <- attach_meta_partner(df)}
}
