% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_meta.R
\name{attach_meta_pepfar_ip}
\alias{attach_meta_pepfar_ip}
\title{Retreive and attach partner metadata}
\usage{
attach_meta_pepfar_ip(df)
}
\arguments{
\item{df}{Dataframe cleaned via reshape_sisma}
}
\value{
A dataframe containing partner metadata
}
\description{
\code{attach_meta_pepfar_ip} Une informações sobre o apoio dos parceiros PEPFAR gravadas no pacote \code{sismar} ao objeto de dados introduzido pelo utilizador
}
\examples{
 \dontrun{

 df <- attach_meta_pepfar_ip(df)}
}
