% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_meta_uid.R
\name{attach_meta_uid}
\alias{attach_meta_uid}
\title{Retreive and attach province and district uid metadata}
\usage{
attach_meta_uid(df, sheetname = "map_distrito_psnuuid")
}
\arguments{
\item{df}{Dataframe cleaned via reshape_sisma}

\item{sheetname}{Sheet name from the MISAU_Site_Map}
}
\value{
A dataframe containing site metadata
}
\description{
\code{attach_meta_uid} utilizes stored user credentials to access
a GoogleSheet where province and district uid metadata is
stored.
}
\examples{
 \dontrun{

 df <- attach_meta_uid(df)}
}
