% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_exact_match.R
\name{check_exact_match}
\alias{check_exact_match}
\title{Função auxiliar para mapear ficheiro de insumo ao modelo de processamento}
\usage{
check_exact_match(df_vars, template_vars)
}
\arguments{
\item{df_vars}{variáveis de dataframe}

\item{template_vars}{variáveis dos modelos de referência}
}
\value{
matriz de correspondência
}
\description{
Função auxiliar para mapear ficheiro de insumo ao modelo de processamento
}
\examples{
 \dontrun{

 df <- check_exact_match()}
}
