% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_disa_dpi.R
\name{clean_disa_dpi}
\alias{clean_disa_dpi}
\title{Função auxiliar para recodificar o quadro de dados DISA}
\usage{
clean_disa_dpi(df)
}
\arguments{
\item{df}{Quadro de dados semi-processado transmitido através de process_disa_dpi}
}
\value{
Quadro de dados DISA recodificado
}
\description{
Função auxiliar para recodificar o quadro de dados DISA
}
\examples{
 \dontrun{

 df <- clean_disa_dpi()}
}
