% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_sisma_csv.R
\name{clean_sisma_csv}
\alias{clean_sisma_csv}
\title{Clean SISMA .csv exports}
\usage{
clean_sisma_csv(file)
}
\arguments{
\item{file}{The path to a .csv file exported from SISMA that contains data
element or indicator variables running wide}
}
\value{
\code{clean_sisma_csv()} returns an object with 7 columns (period, snu,
psnu, sitename, sisma_uid, indicator, value). Unnecessary and duplicate
columns from the input file have been removed, a period column coded from
the original periodcode, and all remaining numeric columns pivoted long.
}
\description{
\code{clean_sisma_csv()} is a helper function for cleaning a .csv pivot table
exported from SISMA
}
\examples{
 \dontrun{
 df <- clean_sisma_csv()}
}
