% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_sisma_csv.R
\name{clean_sisma_csv}
\alias{clean_sisma_csv}
\title{A general function for cleaning a CSV export from SISMA}
\usage{
clean_sisma_csv(file)
}
\arguments{
\item{file}{CSV input pile from SISMA}
}
\value{
A long format of original CSV export from SISMA
}
\description{
A general function for cleaning a CSV export from SISMA
}
\examples{
 \dontrun{

 df <- clean_sisma_csv()}
}
