% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_sisma_csv.R
\name{clean_sisma_csv}
\alias{clean_sisma_csv}
\title{Clean SISMA .csv exports}
\usage{
clean_sisma_csv(file)
}
\arguments{
\item{file}{Path to a .csv file exported from SISMA that contains
data element variables running wide and period and organizational
units running long}
}
\value{
\code{clean_sisma_csv()} returns an object with 7 columns (period, snu,
psnu, sitename, sisma_uid, indicator, value). Unnecessary columns from
the input file are removed and a period column coded from the input
variable periodcode is created. All numeric data element columns that
represent indicator or data element values are pivoted long
}
\description{
\code{clean_sisma_csv()} A helper function that cleans .csv pivot tables
exported from SISMA
}
\examples{
 \dontrun{
 df <- clean_sisma_csv()}
}
