% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_sisma.R
\name{clean_sisma_csv}
\alias{clean_sisma_csv}
\title{Importar e limpar ficheiro .csv do SISMA}
\usage{
clean_sisma_csv(file)
}
\arguments{
\item{file}{Caminho para um ficheiro .csv exportado do SISMA que contém variáveis de elementos no horizontal e período e unidades organizacionais no vertical}
}
\value{
\code{clean_sisma_csv()} devolve um objeto de dados limpo com 7 colunas (período, província, distrito, us, sisma_uid, indicador e valor)
}
\description{
\code{clean_sisma_csv()} importa um ficheiro .csv bruto de SISMA e devolve um objecto transformado numa estrutura longa, cria uma coluna de período formatada e devolve um objeto de dados com 7 colunas (período, província, distrito, us, sisma_uid, indicador e valor)
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
\code{clean_sisma_csv} foi substituída por \code{clean_sisma_df}
}
\examples{
 \dontrun{
 df <- clean_sisma_csv()}
}
