% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_sisma_df.R
\name{clean_sisma_df}
\alias{clean_sisma_df}
\title{Clean SISMA input file prior to parsing}
\usage{
clean_sisma_df(df)
}
\arguments{
\item{df}{}
}
\value{
Cleaned SISMA dataframe
}
\description{
Clean SISMA input file prior to parsing
}
\examples{
 \dontrun{

 df <- clean_sisma_df()}
}
