% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_disa_uid_map}
\alias{data_disa_uid_map}
\title{Mapa de UID para ligação entre DISA, SISMA e DATIM}
\format{
Uma lista das variáveis incluídas no objecto de dados
\describe{
\item{disa_uid}{UID DISA (e.g. CHAEM)}
\item{sisma_uid}{UID SISMA (e.g. x9jhds3ie6)}
\item{datim_uid}{UID DATIM (e.g. jIem8wI1Rx)}
}
}
\usage{
data(data_disa_uid_map)
}
\description{
Mapa de UID para ligação entre DISA, SISMA e DATIM
}
\keyword{datasets}
