% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_sisma_geo_above_site}
\alias{data_sisma_geo_above_site}
\title{Mapa para ligar os códigos de forma da província e do distrito aos dados SISMA}
\format{
Uma lista das variáveis incluídas no objecto de dados
\describe{
\item{provincia}{Provincia}
\item{snuuid}{UID da provincia}
\item{distrito}{Distrito}
\item{psnuuid}{UID do distrito}
\item{distrito_ine}{Distrito no ficheiro INE}
}
}
\usage{
data(data_sisma_geo_above_site)
}
\description{
Mapa para ligar os códigos de forma da província e do distrito aos dados SISMA
}
\keyword{datasets}
