% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_sisma_geo_sites}
\alias{data_sisma_geo_sites}
\title{Map for linking geocoordinates to SISMA health facility data}
\format{
A list of the variables included in the site list
\describe{
\item{data_sisma_geo_sites}{sisma_uid}
\item{data_sisma_geo_sites}{latitude}
\item{data_sisma_geo_sites}{longitude}
}
}
\usage{
data(data_sisma_geo_sites)
}
\description{
List of indicators
}
\keyword{datasets}
