% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_sisma_hiv_prep_map}
\alias{data_sisma_hiv_prep_map}
\title{SISMA PREP Map for Feature Engineering}
\format{
A list of the variables included to engineer dataset features
\describe{
\item{data_sisma_hiv_prep_map}{indicator}
\item{data_sisma_hiv_prep_map}{indicator_new}
\item{data_sisma_hiv_prep_map}{age}
\item{data_sisma_hiv_prep_map}{sex}
\item{data_sisma_hiv_prep_map}{disaggregate}
}
}
\usage{
data(data_sisma_hiv_prep_map)
}
\description{
List of indicators
}
\keyword{datasets}
