% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_sisma_sitelist}
\alias{data_sisma_sitelist}
\title{Deduplicated SISMA site list}
\format{
A list of the variables included in the site list
\describe{
\item{data_sisma_sitelist}{sisma_uid}
\item{data_sisma_sitelist}{provincia}
\item{data_sisma_sitelist}{distrito}
\item{data_sisma_sitelist}{us}
}
}
\usage{
data(data_sisma_sitelist)
}
\description{
List of indicators
}
\keyword{datasets}
