% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_check_disa.R
\name{id_check_disa}
\alias{id_check_disa}
\title{Detetar casos problemáticos de mapeamento de UID DISA para investigação}
\usage{
id_check_disa(df, period_window = 6)
}
\arguments{
\item{df}{Dataframe já processado por \code{process_disa_cv} ou \code{process_disa_dpi}}

\item{period_window}{Número de meses analisados do dataframe providenciado (contado a partir do mês mais recente)}
}
\value{
Dataframe com 1) UID DISA sem UID SISMA correspondentes; e 2) UID SISMA com múltiplos UID DISA
}
\description{
Detetar casos problemáticos de mapeamento de UID DISA para investigação
}
\examples{
 \dontrun{

 df <- id_check_disa()}
}
