% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pop.R
\name{load_pop_ine}
\alias{load_pop_ine}
\title{Esta função lê todos os conjuntos de dados da população de uma pasta especificada}
\usage{
load_pop_ine(filename, input_tabs)
}
\arguments{
\item{filename}{utiliza um caminho fornecido pelo utilizador}

\item{input_tabs}{folhas no ficheiro a ser processado}
}
\value{
um conjunto de dados de todos os ficheiros e separadores
}
\description{
Esta função lê todos os conjuntos de dados da população de uma pasta especificada
}
\examples{
 \dontrun{

 df <- load_pop_ine()}
}
