% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sisma_ats_results.R
\name{parse_sisma_ats_results}
\alias{parse_sisma_ats_results}
\title{Create tidy dataframes for HIV testing services reported via ATS register}
\usage{
parse_sisma_ats_results(df)
}
\arguments{
\item{df}{A dataframe object output from \code{clean_sisma_csv}}
}
\value{
\code{parse_sisma_ats_results} returns a tidy object with 16 columns of
site metadata, indicator features and results
}
\description{
\code{parse_sisma_ats_results} produces a tidy dataframe from an object passed in by
\code{sisma_clean_csv}. It engineers useful data features such as sex, age,
indicator disaggregation, sub-group type, etc.
}
\examples{
 \dontrun{

 df <- parse_sisma_ats_results()}
}
