% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_sisma.R
\name{parse_sisma_ats_saaj_cm}
\alias{parse_sisma_ats_saaj_cm}
\title{Arrumar quadros de dados para os serviços de testagem do HIV dos Serviços
Amigos dos Adolescentes e Jovens (SAAJ)}
\usage{
parse_sisma_ats_saaj_cm(df)
}
\arguments{
\item{df}{Objecto relatório padrão introduzido por \code{process_sisma_export}
ou \code{process_sisma_csv}}
}
\value{
\code{parse_sisma_ats_saaj_cm} devolve um quadro de dados arrumado com 16
colunas de metadados da us, desagregações do indicador, e valores
}
\description{
\code{parse_sisma_ats_saaj_cm} Uma função auxiliar que  gera um quadro de dados arrumado
a partir  de um objecto relatório padrão introduzido por \code{process_sisma_export}
ou \code{process_sisma_csv}. Variáveis úteis que são criados por esta função incluem
sexo, idade, tipo de subgrupo, e outras desagregações presentes na fonte.
}
\examples{
 \dontrun{

 df <- parse_sisma_ats_saaj_cm()}
}
