% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_sisma.R
\name{parse_sisma_csv}
\alias{parse_sisma_csv}
\title{Arrumar quadros de dados SISMA por tipo especificado pelo utilizador}
\usage{
parse_sisma_csv(data, type)
}
\arguments{
\item{data}{Caminho para o ficheiro de dados SISMA (guardado em formato .csv)}

\item{type}{Define o tipo de dados programáticos para processamento}
}
\value{
\code{parse_sisma_csv} devolve um quadro de dados arrumado com 16
colunas de metadados da us, desagregações do indicador, e valores
}
\description{
\code{parse_sisma_csv} função auxiliar que chama a função apropriada
com base no argumento definido pelo utilizador
\subsection{Input File Types}{
\itemize{
\item SMI-CPN: CPN, Saude Materna Infantil
\item SMI-MAT: Maternidade, Saude Materna Infantil
\item SMI-CCR: CCR, Saude Materna Infantil
\item SMI-CPP: CCP, Saude Materna Infantil
\item SMI-CCD: CCD, Saude Materna Infantil
\item SMI-CCS: CCS, Saude Materna Infantil
\item SMI-UG: UG, Saude Materna Infantil
\item SMI-PF: PF, Saude Materna Infantil
\item SMI-PAV: PAV, Saude Materna Infantil
\item ATS Result: ATS Resultados, Programa de HIV
\item ATS History: ATS Historial e Populacoes Chave, Programa de HIV
\item ATS CI: ATS Caso Indice e Ligacao, Programa de HIV
\item ATS SAAJ: ATS Serviço Amigo Adolescente Joven
\item ATS CCSD: Consulta da Crianca Sadia e Consulta da Crianca Doente
\item ATS SMI: ATS especifico a SMI (outros acomopanhantes na CPN, etc.)data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABIAAAASCAYAAABWzo5XAAAAWElEQVR42mNgGPTAxsZmJsVqQApgmGw1yApwKcQiT7phRBuCzzCSDSHGMKINIeDNmWQlA2IigKJwIssQkHdINgxfmBBtGDEBS3KCxBc7pMQgMYE5c/AXPwAwSX4lV3pTWwAAAABJRU5ErkJggg==
\item ATS Auto: ATS Autotestagem, Programa de HIV
\item HIV TARV: Tratamento Antiretroviral (TARV), Programa de HIV
\item HIV PREP: Profilaxia Pré-Exposição (PrEP), Programa de HIV
\item HIV APSS: Apoio Psicossocial (APSS), Programa de HIV
\item HIV ITS: Infecções de Transmissão Sexual (ITS), Programa de HIV
\item HIV AJMHCMM: Adolescent Joven Mentor, Homen Campeao, Maes Mentora
\item HIV DAH: Doenca Avancada do HIV
}
}
}
\examples{
 \dontrun{

 df <- parse_sisma_csv()}
}
