% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sisma_csv.R
\name{parse_sisma_csv}
\alias{parse_sisma_csv}
\title{Cleans and transforms SISMA programmatic data sets into a
tidy format useful for analysis in R/PowerBI/Tableau.}
\usage{
parse_sisma_csv(data, type)
}
\arguments{
\item{data}{Path to SISMA dataset (saved in .csv format)}

\item{type}{Defines programmatic data type for processing}
}
\value{
A tidy dataframe with monthly sisma results
}
\description{
parse_sisma_csv cleans and transforms programmatic
data sets exported from SISMA into tidy data frames. Such
structured data sets are key to facilitating subsequent analysis
in R/PowerBI/Tableau.
\subsection{Overview of programmatic data type options}{
\itemize{
\item SMI-CPN: CPN, Saude Materna Infantil
\item SMI-MAT: Maternidade, Saude Materna Infantil
\item SMI-CCR: CCR, Saude Materna Infantil
\item ATS Result: ATS Resultados, Programa de HIV
\item ATS History: ATS Historial e Populacoes Chave, Programa de HIV
\item ATS CI: ATS Caso Indice e Ligacao, Programa de HIV
\item ATS SAAJ: ATS Serviço Amigo Adolescente Joven
\item ATS CCSD: Consulta da Crianca Sadia e Consulta da Crianca Doente
\item ATS SMI: ATS especifico a SMI (outros acomopanhantes na CPN, etc.)
\item ATS Auto: ATS Autotestagem, Programa de HIV
\item HIV TARV: Tratamento Antiretroviral (TARV), Programa de HIV
\item HIV PREP: Profilaxia Pré-Exposição (PrEP), Programa de HIV
\item HIV APSS: Apoio Psicossocial (APSS), Programa de HIV
}
}
}
\examples{
 \dontrun{

 df <- parse_sisma_csv()}
}
