% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_sisma.R
\name{parse_sisma_hiv_dah}
\alias{parse_sisma_hiv_dah}
\title{Arrumar quadros de dados para os serviços doença avançada do HIV}
\usage{
parse_sisma_hiv_dah(df)
}
\arguments{
\item{df}{Objecto relatório padrão introduzido por \code{process_sisma_export}
ou \code{process_sisma_csv}}
}
\value{
\code{parse_sisma_hiv_dah} devolve um quadro de dados arrumado com 16
colunas de metadados da us, desagregações do indicador, e valores
}
\description{
\code{parse_sisma_hiv_dah} Uma função auxiliar que  gera um quadro de dados arrumado
a partir  de um objecto relatório padrão introduzido por \code{process_sisma_export}
ou \code{process_sisma_csv}. Variáveis úteis que são criados por esta função incluem
sexo, idade, tipo de subgrupo, e outras desagregações presentes na fonte.
}
\examples{
 \dontrun{

 df <- parse_sisma_hiv_dah()}
}
