% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_disa_cv.R
\name{process_disa_cv}
\alias{process_disa_cv}
\title{Processar: Carga Viral (CV) da DISA OpenLDR}
\usage{
process_disa_cv(file, month)
}
\arguments{
\item{file}{Caminho do ficheiro CV .xlsx}

\item{month}{Mês dos dados CV (formatado “AAAA-MM-DD”)}
}
\value{
Um quadro de dados arrumado dos resultados do CV
}
\description{
Processar: Carga Viral (CV) da DISA OpenLDR
}
\examples{
 \dontrun{

 df <- process_disa_cv()}
}
