% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_disa_dpi.R
\name{process_disa_dpi}
\alias{process_disa_dpi}
\title{Tidy DISA DPI .xlsx files (extracted from OpenLDR)}
\usage{
process_disa_dpi(file, period, type = "new")
}
\arguments{
\item{file}{Path of DISA DPI .xlsx file}

\item{period}{Month of DISA DPI extract (passed as "\%Y-\%m-\%d")}

\item{type}{Type of DISA data file to tidy ("old" or "new")}
}
\value{
A tidy dataframe of DISA DPI results
}
\description{
Tidy DISA DPI .xlsx files (extracted from OpenLDR)
}
\examples{
 \dontrun{

 df <- process_disa_dpi()}
}
