% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pop.R
\name{process_pop_ine}
\alias{process_pop_ine}
\title{Title}
\usage{
process_pop_ine(file_inventory, input_sheets, output_type = "MISAU")
}
\arguments{
\item{file_inventory}{xxx}

\item{input_sheets}{yyy}

\item{output_type}{zzz}
}
\value{
df
}
\description{
Title
}
\examples{
 \dontrun{

 df <- process_pop_ine()}
}
