% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_sisma.R
\name{process_sisma_csv}
\alias{process_sisma_csv}
\title{Processar: Dados em formato .csv do SISMA}
\usage{
process_sisma_csv(file, type, language = "portuguese")
}
\arguments{
\item{file}{Caminho do ficheiro csv do sisma}

\item{type}{Tipo de exportação sisma csv (SMI-CPN, Resultado ATS, etc.)}

\item{language}{Lingua dos nomes das variáveis geradas (português ou inglês)}
}
\value{
Um quadro de dados arrumado dos resultados SISMA
}
\description{
\code{process_sisma_csv} é uma função que limpa e transforma as exportações tabulares do SISMA guardadas localmente no formato .csv
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
\code{process_sisma_csv} foi substituído a favor de \code{process_sisma_export}
\subsection{Tipos de ficheiros de entrada}{
\itemize{
\item SMI-CPN: CPN, Saude Materna Infantil
\item SMI-MAT: Maternidade, Saude Materna Infantil
\item SMI-CCR: CCR, Saude Materna Infantil
\item SMI-CPP: CCP, Saude Materna Infantil
\item SMI-CCD: CCD, Saude Materna Infantil
\item SMI-CCS: CCS, Saude Materna Infantil
\item SMI-UG: UG, Saude Materna Infantil
\item SMI-PAV: PAV, Saude Materna Infantil
\item ATS Result: ATS Resultados, Programa de HIV
\item ATS History: ATS Historial e Populacoes Chave, Programa de HIV
\item ATS CI: ATS Caso Indice e Ligacao, Programa de HIV
\item ATS SAAJ: ATS Serviço Amigo Adolescente Joven
\item ATS CCSD: Consulta da Crianca Sadia e Consulta da Crianca Doente
\item ATS SMI: ATS especifico a SMI (outros acomopanhantes na CPN, etc.)
\item ATS Auto: ATS Autotestagem, Programa de HIV
\item HIV TARV: Tratamento Antiretroviral (TARV), Programa de HIV
\item HIV PREP: Profilaxia Pré-Exposição (PrEP), Programa de HIV
\item HIV APSS: Apoio Psicossocial (APSS), Programa de HIV
\item HIV ITS: Infecções de Transmissão Sexual (ITS), Programa de HIV
\item HIV AJMHCMM: Adolescent Joven Mentor, Homen Campeao, Maes Mentora
\item HIV DAH: Doenca Avancada do HIV
}
}
}
\examples{
 \dontrun{

 df <- process_sisma_csv()}
}
