% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_sisma_csv.R
\name{process_sisma_csv}
\alias{process_sisma_csv}
\title{Function that cleans sisma csv exports and parses them into tidy dataframes}
\usage{
process_sisma_csv(file, type, language = "portuguese")
}
\arguments{
\item{file}{Path of sisma csv input}

\item{type}{Type of sisma csv export (SMI-CPN, ATS-Result, etc.)}

\item{language}{Language of variable names (Portuguese or English)}
}
\value{
A tidy data frame of sisma program results
}
\description{
process_sisma_csv cleans and transforms programmatic
data sets exported in CSV format from SISMA into tidy data frames.
Such structured data sets are key to facilitating subsequent analysis
in R/PowerBI/Tableau.
\subsection{Overview of programmatic data type options}{
\itemize{
\item SMI-CPN: CPN, Saude Materna Infantil
\item SMI-MAT: Maternidade, Saude Materna Infantil
\item SMI-CCR: CCR, Saude Materna Infantil
\item ATS Result: ATS Resultados, Programa de HIV
\item ATS History: ATS Historial e Populacoes Chave, Programa de HIV
\item ATS CI: ATS Caso Indice e Ligacao, Programa de HIV
\item ATS SAAJ: ATS Serviço Amigo Adolescente Joven
\item ATS CCSD: Consulta da Crianca Sadia e Consulta da Crianca Doente
\item ATS SMI: ATS especifico a SMI (outros acomopanhantes na CPN, etc.)
\item ATS Auto: ATS Autotestagem, Programa de HIV
\item HIV TARV: Tratamento Antiretroviral (TARV), Programa de HIV
\item HIV PREP: Profilaxia Pré-Exposição (PrEP), Programa de HIV
\item HIV APSS: Apoio Psicossocial (APSS), Programa de HIV
}
}
}
\examples{
 \dontrun{

 df <- process_sisma_csv()}
}
