% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_sisma.R
\name{process_sisma_export}
\alias{process_sisma_export}
\title{Process: Dados em formato .csv do SISMA}
\usage{
process_sisma_export(filename, language = "portuguese")
}
\arguments{
\item{filename}{Caminho do ficheiro .csv exportado do SISMA}

\item{language}{Linga dos nomes das variáveis no quadro gerado (português ou inglês)}
}
\value{
devolve um quadro de dados arrumado com 16 colunas de metadados da us, desagregações do indicador, e valores
}
\description{
\code{process_sisma_export()} utiliza um caminho fornecido pelo utilizador para uma exportação SISMA csv em bruto e executa várias acções de processamento, incluindo arrumação e criação de caraterísticas de dados.
}
\examples{
 \dontrun{

 df <- process_sisma_export()}
}
