% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_sisma.R
\name{process_sisma_export}
\alias{process_sisma_export}
\title{Arrumar quadros de dados a partir de relatórios padrão SISMA em formato .csv}
\usage{
process_sisma_export(filename, language = "portuguese")
}
\arguments{
\item{filename}{Caminho do ficheiro .csv exportado do SISMA}

\item{language}{Linga dos nomes das variáveis no quadro gerado (português ou inglês)}
}
\value{
devolve um quadro de dados arrumado com 16 colunas de metadados da us,
desagregações do indicador, e valores
}
\description{
Arrumar quadros de dados a partir de relatórios padrão SISMA em formato .csv
}
\examples{
 \dontrun{

 df <- process_sisma_export()}
}
