% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_example_csv.R
\name{read_example_csv}
\alias{read_example_csv}
\title{Read example SISMA CSV data into environment}
\usage{
read_example_csv(example_data)
}
\arguments{
\item{example_data}{The name of the example data file to read}
}
\value{
A data frame with the example data
}
\description{
Read example SISMA CSV data into environment
}
\examples{
 \dontrun{

 df <- example_data()}
}
