% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_vars.R
\name{seq_vars}
\alias{seq_vars}
\title{Função auxiliar que reordena variáveis em sequência padrão}
\usage{
seq_vars(df)
}
\arguments{
\item{df}{Objecto quadro de dados contendo todos variáveis padrão}
}
\value{
devolve um quadro de dados arrumado com 16
variáveis reordenados
}
\description{
Função auxiliar que reordena variáveis em sequência padrão
}
\examples{
 \dontrun{

 df <- seq_vars()}
}
