% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_language.R
\name{set_language}
\alias{set_language}
\title{Define a lingua dos variáveis no quadro gerado}
\usage{
set_language(df, language = "portuguese")
}
\arguments{
\item{df}{Objecto relatório padrão introduzido por \code{process_sisma_export}
ou \code{process_sisma_csv}}

\item{language}{Língua desejada ("portuguese" ou "english")}
}
\value{
Quadro de dados arrumado com nomes das variáveis na lingua desejada
}
\description{
Define a lingua dos variáveis no quadro gerado
}
\examples{
 \dontrun{

 df <- set_language()}
}
