% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_sismar_folder.R
\name{setup_sismar_folder}
\alias{setup_sismar_folder}
\title{Cria uma estrutura organizacional (pastas) que é comum a todos os projectos sismar}
\usage{
setup_sismar_folder(
  folder_list = list("Data", "Data/sisma", "Data/disa", "Data/pepfar", "Data/ine",
    "Data/sismar/processed", "Data/disa/processed", "Data/pepfar/processed", "Images",
    "Scripts", "AI", "Dataout", "Data_public", "GIS", "Documents", "Graphics",
    "markdown")
)
}
\arguments{
\item{folder_list}{lista de pastas a instalar}
}
\value{
Um conjunto padrão de pastas no projeto para utilizar com as funções \code{sismar}
}
\description{
Cria uma estrutura organizacional (pastas) que é comum a todos os projectos sismar
}
\examples{
 \dontrun{

 setup_sismar_folder()}
}
