% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sismar-package.R
\docType{package}
\name{sismar-package}
\alias{sismar}
\alias{sismar-package}
\title{sismar: Transform tabular export files from SISMA}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a set of functions for creating analytical datasets from SISMA CSV downloads. This includes functions that reshape files into long format, remove unnecessary variables, and engineer features useful in analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://usaid-mozambique.github.io/sismar/}
  \item Report bugs at \url{https://github.com/usaid-mozambique/sismar/issues}
}

}
\author{
\strong{Maintainer}: Joe Lara \email{jlara@usaid.gov}

Authors:
\itemize{
  \item Karishma Srikanth \email{ksrikanth@usaid.gov}
  \item Kirstin Lyon \email{kirstin.c.lyon@gmail.com}
}

}
\keyword{internal}
